@@ -56,19 +48,46 @@
   "author": "Helder Santana",
   "license": "MIT",
   "devDependencies": {
-    "babel": "^5.8.23",
+    "babel-cli": "^6.1.2",
+    "babel-core": "^6.1.2",
+    "babel-polyfill": "^6.0.16",
+    "babel-preset-es2015": "^6.1.2",
+    "babel-preset-es2015-node4": "^1.0.0",
+    "babel-preset-stage-0": "^6.1.2",
     "blanket": "~1.1.6",
-    "chai": "~1.9.0",
-    "coveralls": "~2.8.0",
-    "eslint": "^1.6.0",
-    "mocha": "^1.21.5",
-    "mocha-lcov-reporter": "0.0.1",
+    "chai": "~3.4.0",
+    "coveralls": "~2.11.4",
+    "eslint": "^1.9.0",
+    "image-size": "^0.3.5",
+    "mimer": "^0.2.1",
+    "minimist": "^1.2.0",
+    "mocha": "^2.3.3",
+    "mocha-lcov-reporter": "1.0.0",
     "semver": "^5.0.3",
-    "sinon": "~1.8.1"
+    "sinon": "~1.17.2"
   },
   "config": {
     "blanket": {
-      "pattern": "datauri/dist/"
+      "pattern": "datauri/lib/"
+    }
+  },
+  "datauri-build": {
+    "datauri": {
+      "dependencies": [
+        "image-size",
+        "mimer",
+        "semver"
+      ],
+      "description": "Simple DataURI scheme module"
+    },
+    "datauri-cli": {
+      "dependencies": [
+        "minimist"
+      ],
+      "bin": {
+        "datauri": "./index.js"
+      },
+      "description": "Simple DataURI scheme client"
     }
   }
 }
