@@ -1,4 +1,4 @@
-import Api from './datauri/api';
+import Api from './api';
 
 const ENC_TYPE = configSize => configSize > 1 ? 'encode' : 'encodeSync';
 
